Create Length_of_longest_palindromic_substring(DP + loops).cpp
Complexity Analysis:  

Time complexity: O(n^2). 
Two nested traversals are needed.
Auxiliary Space: O(n^2). 
Matrix of size n*n is needed to store the dp array.



METHOD 3: Using loops

APPROACH: First we will run a loop for iterating every character. Then we will run another loop inside it to check that is there any other character similar to the current character. If it is, then it is possible that they both are first and last character of longest substring. We will store that substring and check whether that substring is longest or not. If yes then we will store that substring and keep iterating.


// A C++ solution for longest palindrome
#include <bits/stdc++.h>
using namespace std;
// This function prints the
// longest palindrome substring
// It also returns the length
// of the longest palindrome
int longestPalSubstr(string str)
{
    //Stores Longest Pallidrome Substring
    string longest = "";
    int n = str.length();
    int j;
    //To store substring which we think can be a pallindrome
    string subs = "";
    //To strore reverse of substring we think can be pallidrome
    string subsrev = "";
    for(int i = 0; i < n; i++){
        j = n-1;
        while(i < j){
            //Checking whether the character at i and j are same. If they are same then that substring can be LPS
            if((str[i] == str[j]) && (longest.length() < (j-i+1))){
               subs = str.substr(i,(j-i+1));
               //cout<<subs<<" ";
               subsrev = subs;
               reverse(subsrev.begin(), subsrev.end());
               if(subs == subsrev){
                  longest = subs;
               }
            }
            j--;
        }
    }
   //If no longest substring then we will return first character(In Leetcode it was a testcase so...)
    if(longest.length() == 0){
        longest = str[0];
    }
    cout << "Longest palindrome substring is: " << longest;
 
    // return length of LPS
    return longest.length();
}
 
// Driver Code
int main()
{
    string str = "forgeeksskeegfor";
    cout << "\nLength is: "
        << longestPalSubstr(str);
    return 0;
}
Output
Longest palindrome substring is: geeksskeeg
Length is: 10
Time and Space Complexity Analysis:

Time Complexity – Worst case -> O(n^3) because used 2 nested loops and a reverse function in 2nd nested loop.

But better than first solution because we were checking palindrome for all substrings but here we are checking for some cases only.

Auxiliary Space – O(3*(length of longest palindrome))  – To store answer, substring which we think can be palindrome and reverse of string which we think is a palindrome.