Create Shortest path with Alternating Colors(Leetcode).cpp
Objective: The goal here is to find the shortest path between the source node to every other node in the graph with an additional condition that the path between any two nodes should be color alternating.

Intuition: We already know that we can find the shortest path between any two nodes in a graph using BFS.
The Brute force Idea is to use BFS traversal n times so that we could find out the shortest path between all the nodes from the source node 0.
But doing this is an overkill. We can acheive this using a single BFS traversal.

Approach: Since we can visit all nodes in a graph using BFS traversal, we can update the shortest distance for every node when visiting that particular node. The shortest distance between source node and target node is the distance when we first visit the target node starting from source.

Algorithm:

First, we construct an adjacency matrix with the given red_edges and blue_edges.
It is important for us to keep in mind the fact that there can be at most two directed edges(in the same direction) between any two nodes(a blue edge and a red edge )
Here I am using a pair of int and int to keep track of the node index and the color of the edge respectively ( 0 for a red edge and 1 for a blue edge)
I am keeping track of three values in the queue, the first one to denote the node index, the second one for the distance and the third one for color ( the color is -1 only for the source node)
Use a BFS traversal and keep updating the dist only when it is not visited. Here I am updating the node value in the adjacency matrix to -1 to denote that the node has already been visited.