Create Update_queries(GFG).cpp
If you have multiple queries telling to increment the numbers in the range l to r, then you have to just increment number at l and decrement number at r+1 and after performing all queries you just do prefix sum for all indices.

Same idea is to be used for OR queries. So to convert OR to Range sum we will use log2(1e5)~17 bits initialized to 0.And for each query we will increment all the set bits of the number to be incremented(here 3rd number in the vector) at l and decrement all the set bits at r+1. Similarly perform prefix sum at each bit of the index and we will have the final answer for that index.