Create Sum Room to Leaf numbers(Leetcode).cpp
We are given a binary tree consisting of nodes with digit values. We need to consider every possible root-to-leaf path. Each of these path represent a number. We need to return the total sum of all the number represented by each of these paths.

✔️ Solution - I (Recursive DFS - Preorder)

We can intuitively see that a DFS solution will fit well here since we traverse from root to leaf in a DFS traversal and thus we can use it to form number represented by each root-to-leaf path. In DFS traversal -

We start at the root with current number cur = 0.
Every time, the current node's digit will be appended to cur from root to leaf and we recurse for left and right child.
If we reach a leaf node (a node not having any left or right child), we have formed the complete root-to-leaf number. We will then add this number cur to the overall sum.
Sum = 0, Cur = 4	Sum = 0, Cur = 49	Sum = 495, Cur = 495	Sum = 986, Cur = 491	Sum = 1026, Cur = 40
				

Time Complexity : O(N), where N is the number of nodes in the tree. We are doing a standard DFS traversal which takes O(N) time
Space Complexity : O(H), where H is the maximum depth of tree. This space is required for implicit recursive stack space. In the worst case, the tree maybe skewed and H = N in which case space required is equal to O(N).

✔️ Solution - II (Iterative DFS - preorder)

The same as above but converted to iterative solution using an explicit stack. The working and complexities should be about the same for both the solutions. But, an interviewer may ocassionally ask to implement iterative version, so it's good to know this as well.


Time Complexity : O(N), where N is the number of nodes in the tree. We are doing a standard DFS traversal which takes O(N) time
Space Complexity : O(H)

✔️ Solution - III (BFS)

The problem doesn't intuitively fall under BFS but it can be solved using this traversal approach as well.

class Solution {
public:
    int sumNumbers(TreeNode* root) {
        int sum = 0;
        queue<pair<TreeNode*, int>> q;
        q.push({root, 0});
        while(q.size()) {
            auto [n, cur] = q.front(); q.pop();
            cur = cur * 10 + n -> val;
            if(n -> left) q.push({n -> left, cur});      
            if(n -> right) q.push({n -> right, cur});
            if(!n -> left && !n -> right) sum += cur;   // add to total sum if we are at leaf node
        }
        return sum;
    }
};
Time Complexity : O(N), standard time complexity of a BFS. Each node is only visited once.
Space Complexity : O(N), atmost (N+1)/2 elements will be stored in queue in case of complete binary tree.

✔️ Solution - IV (Morris Traversal - Preorder)

This solution involves using the morris traversal technique to solve the problem. The advantage of this traversal is that we can traverse the tree in O(1) space complexity. The basic idea is to link predecessors to root nodes so we can trace it back once we have traversed a side without need of using a stack. It's a bit hard to explain in words so I have attached an image of how it works below. You can also find some good explanation here.