Create Construct a binary tree from inorder and postorder (Leetcode).cpp
// problem statement

Given two integer arrays inorder and postorder where inorder is the inorder traversal of a binary tree and postorder is the postorder traversal of the same tree, construct and return the binary tree.

 

Example 1:


Input: inorder = [9,3,15,20,7], postorder = [9,15,7,20,3]
Output: [3,9,20,null,null,15,7]
Example 2:

Input: inorder = [-1], postorder = [-1]
Output: [-1]
 

Constraints:

1 <= inorder.length <= 3000
postorder.length == inorder.length
-3000 <= inorder[i], postorder[i] <= 3000
inorder and postorder consist of unique values.
Each value of postorder also appears in inorder.
inorder is guaranteed to be the inorder traversal of the tree.
postorder is guaranteed to be the postorder traversal of the tree.


// approach

 Solution - I (Recursive)

First, we need to observe that we cant form exact binary tree with only 1 traversal.

For eg. inorder = [9,3,15,20,7], here we dont know which is the root node of tree and their children. It could start with any node as root and form a valid binary tree.
Similarly, postorder = [9,15,7,20,3], here we know 3 is the root node since traversal ended with it, however we are not sure of its children. It could have been all rest nodes on right side as (root)3 -> 20 -> 7 -> 15 -> 9 or all on left side as 9 <- 15 <- 7 <- 20 <- 3(root) or any intermediate of these.
But, when we consider both traversals, we can form a unique tree. Since the postorder traversal ends with root node, we know the root node of tree is last node of postorder. At the same time, we can search for that node in inorder and we know that the nodes that occur to its left form its left sub-tree and nodes occuring to the right of it form its right sub-tree. We recursively repeat the same process -

Next node in the line from last from postorder forms the root node, we search it in inorder and see what its left and right subtree are, then recurse for right subtree first, then the left one...until all nodes are traversed.
Step

Description

	The red node-3 denotes current root node
The blue nodes on left and right of red node denotes its left & right-subtrees respectively
We recursively construct the right subtree and then left subtree
	We have recursed to form the right subtree of previous root node-3
20 is the next node from end in postorder traversal and so it forms the current node
Again, the blue nodes on left & right of 20 denote its left & right-subtrees respectively
We again recursively construct the right subtree and then left subtree
	We have recursed to form the right subtree of previous root node-20
7 is next node in postorder which forms current node
There are no left and right subtrees for 7
So just create node-7 and return
	Now, we recurse from step-2 to form left subtree of root node-20
Again, 15 is next node in postorder and it forms the current node
Again, there are no left and right subtrees
So just create node-15 and return
	Now, we recurse from step-1 to form left subtree of root node-3
9 is the next node in postorder and it forms current node
We dont have any left and right subtree for 9
So just create node-9 and return
	All the nodes have been traversed
The recursion stops here and we return the final tree
So to summarize the algorithm can be stated as -

We start by initializing postIdx = n-1, where n is the number of nodes. postIdx denotes the index of postorder traversal that we are at. It will give us the root node for each point of recursion
We also initialize inStart = 0 and inEnd = n-1 which denotes the current range in inorder that the subtree will lie in.
At each recursion, we choose postorder[postIdx] as current root node. We search that index of that node in inorder as well. Then all nodes to the left of that index (till the previous root) are part of current root's left sub-tree and all nodes to the right of it are part of its right sub-tree.
We decrement postIdx and recurse to build the right sub-tree and then left sub-tree.
Why right first? Because postorder goes as left->right->root and we are traversing it from end. So we consider in order root->right->left)
The process is continued till all nodes are traversed.
C++

class Solution {
public:
    TreeNode* buildTree(vector<int>& inorder, vector<int>& postorder) {
        int n = size(inorder), postIdx = n-1;
        return build(inorder, postorder, 0, n-1, postIdx);
    }

    TreeNode* build(vector<int>& in, vector<int>& post, int inStart, int inEnd, int& postIdx) {
        if(inStart > inEnd) return nullptr;
        TreeNode* root = new TreeNode(post[postIdx--]);
        int inIdx = find(begin(in), end(in), root -> val) - begin(in);
        root -> right = build(in, post, inIdx+1, inEnd, postIdx);
        root -> left  = build(in, post, inStart, inIdx-1, postIdx);
        return root;
    }
};
Python

Time Complexity : O(N2) where N is the number of nodes in the tree. At each recursion, we are choosing post[postIdx] and searching it in inorder. This may take O(N) for each search and we are building a tree with N such nodes. So total time complexity is O(N2)
Space Complexity : O(N), required by implicit recursive stack. In worst case of skewed tree, the recursion depth might go upto N.

✔️ Solution - II (Recursive w/ Hashmap)

In the previous approach, we had to search a node in the inorder array everytime which cost us O(N) and led to higher time complexity. However, we can reduce this cost if we maintain a mapping of node's value to inorder node index in a hashmap. This allows us to fetch index of a node in inorder array in O(1) time at each recursion.

C++

class Solution {
public:
    unordered_map<int, int> mp;
    TreeNode* buildTree(vector<int>& inorder, vector<int>& postorder) {
        for(int i = 0; i < size(inorder); i++) mp[inorder[i]] = i;
        int n = size(inorder), postIdx = n-1;
        return build(inorder, postorder, 0, n-1, postIdx);
    }

    TreeNode* build(vector<int>& in, vector<int>& post, int inStart, int inEnd, int& postIdx) {
        if(inStart > inEnd) return nullptr;
        TreeNode* root = new TreeNode(post[postIdx--]);
        int inIdx = mp[root -> val];
        root -> right = build(in, post, inIdx+1, inEnd, postIdx);
        root -> left  = build(in, post, inStart, inIdx-1, postIdx);
        return root;
    }
};
Time Complexity : O(N), now we can find inIdx in O(1) time. So the time required to build the tree out of N nodes each taking O(1) comes out to be O(N)
Space Complexity : O(N), requried by recursive stack and to maintain hashmap.